qaa/requirements.txt
Django==2.0.7
django-extensions==2.0.7
djangorestframework==3.8.2
pytz==2018.5
six==1.11.0
Werkzeug==0.14.1
psycopg2==2.7.5
textblob==0.15.1  # nlp